requirements.txt
# Core
aiogram==3.4.0
fastapi==0.115.0
uvicorn==0.30.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0

# Redis & Celery
redis==5.0.1
celery==5.3.4
apscheduler==3.10.4

# Parsing
requests==2.31.0
beautifulsoup4==4.12.0
lxml==5.1.0

# Config & Monitoring
python-dotenv==1.0.0
pydantic>=2.4.1,<2.6
pydantic-settings==2.1.0
prometheus-client==0.19.0

# Optional (for charts)
# matplotlib==3.7.0
# pandas==2.0.0
